+++ src/britive/secrets_manager.py
@@ -1,4 +1,5 @@
 import json
+from .system.policies import SystemPolicies
 
 
 class SecretsManager:
@@ -152,7 +153,7 @@ class PasswordPolicies:
         has_lower_case_chars: bool = True,
         has_numbers: bool = True,
         has_special_chars: bool = True,
-        allowed_special_chars: str = '@#$%\\(\',)'
+        allowed_special_chars: str = '~`!@#$%^&*()-_+=[]{}|/;:"?/\\.><,\''
     ) -> dict:
         """
         Creates a new password policy.
@@ -161,10 +162,10 @@ class PasswordPolicies:
         :param description: description of the password policy
         :param password_type: type of password to use for the policy
         :param min_password_length: minimum length of the password
-        :param has_upper_case_chars: whether or not to require uppercase characters
-        :param has_lower_case_chars: whether or not to require lowercase characters
-        :param has_numbers: whether or not to require numbers
-        :param has_special_chars: whether or not to require special characters
+        :param has_upper_case_chars: whether to require uppercase characters
+        :param has_lower_case_chars: whether to require lowercase characters
+        :param has_numbers: whether to require numbers
+        :param has_special_chars: whether to require special characters
         :param allowed_special_chars: a string of special characters to allow in the password
         :return: Details of the newly created password policy.
         """
@@ -175,19 +176,14 @@ class PasswordPolicies:
             'passwordType': password_type,
             'minPasswordLength': min_password_length,
             'hasUpperCaseChars': has_upper_case_chars,
-            'hasLowercaseChars': has_lower_case_chars,
+            'hasLowerCaseChars': has_lower_case_chars,
             'hasNumbers': has_numbers,
             'hasSpecialChars': has_special_chars,
             'allowedSpecialChars': allowed_special_chars,
         }
         return self.britive.post(self.base_url, json=params)
 
-    def create_pin(
-        self, 
-        name: str, 
-        description: str = 'Default description', 
-        pin_length: int = 4
-    ) -> dict:
+    def create_pin(self,  name: str,  description: str = 'Default description',  pin_length: int = 4) -> dict:
         """
         Creates a new pin password policy.
 
@@ -215,16 +211,18 @@ class PasswordPolicies:
             description: description of the password policy
             passwordType: type of password to use for the policy
             minPasswordLength: minimum length of the password
-            hasUpperCaseChars: whether or not to require uppercase characters
-            hasLowercaseChars: whether or not to require lowercase characters
-            hasNumbers: whether or not to require numbers
-            hasSpecialChars: whether or not to require special characters
+            hasUpperCaseChars: whether to require uppercase characters
+            hasLowercaseChars: whether to require lowercase characters
+            hasNumbers: whether to require numbers
+            hasSpecialChars: whether to require special characters
             allowedSpecialChars: a string of special characters to allow in the password
             pinLength: the length of the pin to use for the policy (only for pins)
         :return: None
         """
 
-        return self.britive.patch(f'{self.base_url}/{password_policy_id}', json=kwargs)
+        current = self.get(password_policy_id=password_policy_id)
+
+        return self.britive.patch(f'{self.base_url}/{password_policy_id}', json={**current, **kwargs})
 
     def delete(self, password_policy_id: str) -> None:
         """
@@ -241,13 +239,11 @@ class PasswordPolicies:
         Generates a password for the given password policy.
 
         :param password_policy_id: the ID of the password policy
-        :return: the generated the generated password
+        :return: the generated password
         """
 
         params = {'action': 'generatePasswordOrPin'}
-        return self.britive.get(f'{self.base_url}/{password_policy_id}', params=params)[
-            'passwordOrPin'
-        ]
+        return self.britive.get(f'{self.base_url}/{password_policy_id}', params=params)['passwordOrPin']
 
     def validate(self, password_policy_id: str, password: str) -> dict:
         """
@@ -255,13 +251,18 @@ class PasswordPolicies:
 
         :param password_policy_id: the ID of the password policy
         :param password: the password to validate
-        :return: whether or not the password is valid
+        :return: whether the password is valid
         """
 
-        params = {'id': password_policy_id, 'passwordOrPin': password}
-        return self.britive.post(
-            f'{self.base_url}?action=validatePasswordOrPin', json=params
-        )
+        data = {
+            'id': password_policy_id,
+            'passwordOrPin': password,
+        }
+
+        params = {
+            'action': 'validatePasswordOrPin'
+        }
+        return self.britive.post(self.base_url, json=data, params=params)
 
 
 class Folders:
@@ -458,72 +459,104 @@ class Policies:
         params = {'consumer': 'secretmanager', 'resource': path}
         return self.britive.delete(f'{self.base_url}/{policy_id}', params=params)
 
-    def build(
-        self,
-        name: str,
-        description: str = '',
-        draft: bool = False,
-        active: bool = True,
-        read_only: bool = False,
-        users: list = None,
-        tags: list = None,
-        service_identities: list = None,
-        ips: list = None,
-        from_time: str = None,
-        to_time: str = None,
-        approval_notification_medium: str = None,
-        time_to_approve: int = 5,
-        approver_users: list = None,
-        approver_tags: list = None,
-        access_type: str = 'Allow',
-    ) -> dict:
-
+    def build(self, name: str, access_level: str = None, description: str = '', draft: bool = False,
+              active: bool = True, read_only: bool = False, users: list = None, tags: list = None, tokens: list = None,
+              service_identities: list = None, ips: list = None, from_time: str = None, to_time: str = None,
+              date_schedule: dict = None, days_schedule: dict = None, approval_notification_medium: str = None,
+              time_to_approve: int = 5, access_validity_time: int = 120, approver_users: list = None,
+              approver_tags: list = None, access_type: str = 'Allow', identifier_type: str = 'name') -> dict:
         """
-        Builds a policy to be uploaded to the vault.
+        Build a policy document given the provided inputs.
 
-        :param name: name of the policy
-        :param description: description of the policy
-        :param draft: whether or not the policy is a draft
-        :param active: whether or not the policy is active
-        :param read_only: whether or not the policy is read only
-        :param users: list of users to apply the policy to
-        :param tags: list of tags to apply the policy to
-        :param service_identities: list of service identities to apply the policy to
-        :param ips: list of IPs to apply the policy to
-        :param from_time: start time of the policy
-        :param to_time: end time of the policy
-        :param approval_notification_medium: notification medium to use for approval
-        :param time_to_approve: time to approve the policy
-        :param approver_users: list of users to approve the policy
-        :param approver_tags: list of tags to approve the policy
-        :param access_type: access type of the policy
-        :return: policy to be uploaded to the vault.
+        :param name: The name of the policy.
+        :param access_level: The level of access. Valid values are SM_View, SM_Manage, SM_CRUD. Defaults to SM_View
+            if not provided.
+        :param description: An optional description of the policy.
+        :param draft: Indicates if the policy is a draft. Defaults to `False`.
+        :param active: Indicates if the policy is active. Defaults to `True`.
+        :param read_only: Indicates if the policy is a read only. Defaults to `False`.
+        :param users: Optional list of user names or ids to which this policy applies.
+        :param tags: Optional list of tag names or ids to which this policy applies.
+        :param tokens: Optional list of token names or ids to which this policy applies.
+        :param service_identities: Optional list of service identity names or ids to which this policy applies.
+        :param ips: Optional list of IP addresses for which this policy applies. Provide in CIDR notation
+            or dotted decimal format for individual (/32) IP addresses.
+        :param from_time: The start date/time of when the policy is in effect. If a date is provided
+            (`YYYY-MM-DD HH:MM:SS`) this will represent the start date/time of 1 contiguous time range. If just a
+            time is provided (`HH:MM:SS`) this will represent the daily recurring start time. If this parameter is
+            provided then `to_time` must also be provided. This parameter is deprecated as of v2.19.0. The presence of
+            `date_schedule` and/or `days_schedule` will override this field.
+        :param to_time: The end date/time of when the policy is in effect. If a date is provided
+            (`YYYY-MM-DD HH:MM:SS`) this will represent the end date/time of 1 contiguous time range. If just a
+            time is provided (`HH:MM:SS`) this will represent the daily recurring end time. If this parameter is
+            provided then `from_time` must also be provided. This parameter is deprecated as of v2.19.0. The presence of
+            `date_schedule` and/or `days_schedule` will override this field.
+        :param date_schedule: A dict in the format
+            {
+                'fromDate': '2022-10-29 10:30:00',
+                'toDate': '2022-11-05 18:30:00',
+                'timezone': 'UTC'
+            }
+            Timezone formats can be found in the TZ Identifier column of the following page.
+            https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+        :param days_schedule: A dict in the format
+            {
+                'fromTime': '10:30:00',
+                'toTime': '18:30:00',
+                'timezone': 'UTC',
+                'days': ['MONDAY','TUESDAY','WEDNESDAY','THURSDAY','FRIDAY','SATURDAY','SUNDAY']
+            }
+            Timezone formats can be found in the TZ Identifier column of the following page.
+            https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+        :param approval_notification_medium: Optional notification medium name to which approval requests will be
+            delivered. Specifying this parameter indicates the desire to enable approvals for this policy.
+        :param time_to_approve: Optional number of minutes to wait for an approval before denying the action. Defaults
+            to 5 minutes.
+        :param access_validity_time: Optional number of minutes the access is valid after approval. Defaults to 120
+            minutes.
+        :param approver_users: Optional list of user names or ids who are to be considered approvers.
+            If `approval_notification_medium` is set then either `approver_users` or `approver_tags` is required.
+        :param approver_tags: Optional list of tag names who are considered approvers.
+            If `approval_notification_medium` is set then either `approver_users` or `approver_tags` is required.
+        :param access_type: The type of access this policy provides. Valid values are `Allow` and `Deny`. Defaults
+            to `Allow`.
+        :param identifier_type: Valid values are `id` or `name`. Defaults to `name`. Represents which type of
+            identifiers are being provided to the other parameters. Either all identifiers must be names or all
+            identifiers must be IDs.
+        :return: A dict which can be provided as a secret manager policy to `create` and `update`.
         """
 
         policy = self.britive.policies.build(
             name=name,
-            active=active,
             description=description,
             draft=draft,
-            access_type=access_type,
+            active=active,
             read_only=read_only,
             users=users,
             tags=tags,
+            tokens=tokens,
             service_identities=service_identities,
             ips=ips,
             from_time=from_time,
             to_time=to_time,
+            date_schedule=date_schedule,
+            days_schedule=days_schedule,
             approval_notification_medium=approval_notification_medium,
             time_to_approve=time_to_approve,
+            access_validity_time=access_validity_time,
             approver_users=approver_users,
             approver_tags=approver_tags,
+            access_type=access_type,
+            identifier_type=identifier_type
         )
+
         policy.pop('permissions', None)
         policy.pop('roles', None)
-        policy['accessLevel'] = 'SM_Manage'
+        policy['accessLevel'] = access_level or 'SM_View'
+        policy['consumer'] = 'secretmanager'
         return policy
 
-    def create(self, policy: dict, path: str = '/') -> dict:
+    def create(self, policy: dict, path: str) -> dict:
         """
         Creates  a policy in the vault.
 
@@ -531,6 +564,8 @@ class Policies:
         :param path: path of the policy, include the / at the beginning
         :return: Details of the policy.
         """
+
+        policy['resource'] = path
 
         return self.britive.post(f'{self.base_url}?resource={path}&consumer=secretmanager', json=policy)
 
@@ -617,8 +652,10 @@ class StaticSecretTemplates:
         :return: None
         """
 
+        current = self.get(secret_template_id=static_secret_template_id)
+
         return self.britive.patch(
-            f'{self.base_url}/{static_secret_template_id}', json=kwargs
+            f'{self.base_url}/{static_secret_template_id}', json={**current, **kwargs}
         )