+++ tests/test_990-delete_all_resources.py
@@ -21,6 +21,9 @@ def test_profile_delete(cached_profile):
         assert cached_profile['papId'] not in [p['papId'] for p in profiles]
     finally:
         cleanup('profile')
+        cleanup('profile-approval-policy')
+        cleanup('dynamic-session-attribute')
+        cleanup('static-session-attribute')
 
 
 # 070-environments
@@ -133,6 +136,17 @@ def test_folder_delete(cached_folder, cached_vault):
         cleanup('folder')
 
 
+def test_secret_delete(cached_secret, cached_vault):
+    try:
+        response = britive.secrets_manager.secrets.delete(
+            path=cached_secret['path'],
+            vault_id=cached_vault['id']
+            )
+        assert response is None
+    finally:
+        cleanup('secret')
+
+
 def test_static_secret_templates_delete(cached_static_secret_template):
     try:
         response = britive.secrets_manager.static_secret_templates.delete(
@@ -156,17 +170,6 @@ def test_pinPolicy_delete(cached_PinPolicies):
         assert response is None
     finally:
         cleanup('pin-policies')
-
-
-def test_secret_delete(cached_secret, cached_vault):
-    try:
-        response = britive.secrets_manager.secrets.delete(
-            path=cached_secret['path'],
-            vault_id=cached_vault['id']
-            )
-        assert response is None
-    finally:
-        cleanup('secret')
 
 
 def test_policy_delete(cached_policy):